Use perma id for audio hotspot files
When creating a new area or looking up the target of an area, audio
files were referenced by id instead of perma id. After creating the
area, the file input was therefore empty. Selecting a file via the
file input caused the area list thumbnail to display a missing
indicator instead of the pictogram.

REDMINE-17787