Fix Zeitwerk compatibility
Add paths to both `autoload_paths` and
`eager_load_paths`. `eager_load_paths` is required to have paths
covered by `zeitwerk:check`.

Ignore root file in lib which does not follow Zeitwerk conventions,
but is required for Bundler to be able to require the gem.

Ignore and require Paperclip processors which are cached globally and
can, thus, not be reloaded.

For Rails < 6, we cannot add `lib` to the `eager_load_paths` since
Resque eager loads [1], which then evaluates the engine class a second
time. This time `autoload_paths` is already frozen, resulting in an
error.

REDMINE-19438

[1] https://github.com/resque/resque/blob/v1.27.4/lib/resque/tasks.rb#L45